move email module out of auth module
